refactor(core): consolidate 3 telemetry singletons → telemetry-runtime.js
telemetry-singleton, usage-rollup-singleton, and plugin-subscribers-singleton
were one composition concern ("build this process's telemetry runtime") smeared
across three module caches — three loadAppDefinition() try/catches and three
reset*ForTests() that tests had to reset together or leave the rollup subscribed
to a dead bus. Collapse into a single telemetry-runtime.js with one shared state,
per-part lazy getters, and one atomic reset. Bus identity across the parts is now
guaranteed by construction (rollup + plugin subscribers attach to the same
getTelemetry() reached through the module's own closure). Same getter names, so
create-handler's wiring is unchanged; the three reset seams become
resetTelemetryRuntimeForTests + setTelemetryForTests.
